chore(kineokit): separate commands into different files
